[ADD+FIX] reorganized Picking List views and field for multiple parsers
