+++ stock_external_parser_springer/__openerp__.py
@@ -41,6 +41,7 @@ Tasks:
         'views/product_magazine_view.xml',
         'views/file_view.xml',
         'views/res_partner_view.xml',
+        'views/stock_picking_view.xml',
 
         'data/ir_cron.xml',
         # 'security/ir.model.access.csv',
