+++ stock_external_parser_springer/models/__init__.py
@@ -3,3 +3,4 @@ from . import springer_ftp_settings
 from . import springer_file
 from . import springer_job
 from . import springer_product
+from . import springer_stock_picking
